+++ src/main/java/com/greedy/mokkoji/db/review/repository/ReviewRepository.java
@@ -1,6 +1,6 @@
-package com.greedy.mokkoji.db.report.repository;
+package com.greedy.mokkoji.db.review.repository;
 
-import com.greedy.mokkoji.db.report.entity.Review;
+import com.greedy.mokkoji.db.review.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
